requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
idna==2.8
Markdown==3.1
Pillow==6.0.0
pkg-resources==0.0.0
pytz==2018.9
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.1
